+++ app/plugins/auth.ts
@@ -1,3 +1,5 @@
+import { useAuthStore } from "~/stores/auth"
+
 export default defineNuxtPlugin(async () => {
   const authStore = useAuthStore()
   await authStore.fetchUser()
