+++ server/services/users/loginService.ts
@@ -1,6 +1,6 @@
-import type { User } from "~/app/types/auth"
-import { setToken } from "~/server/repositories/user.repository"
-import { getUserByPhone, createUser } from "~/server/repositories/user.repository"
+import type { User } from "~~/shared/types/auth"
+import { setToken } from "~~/server/repositories/user.repository"
+import { getUserByPhone, createUser } from "~~/server/repositories/user.repository"
 
 export const loginService = {
   /**
