+++ server/services/users/userService.ts
@@ -1,5 +1,5 @@
-import { getUserById } from '../../repositories/user.repository'
-import type { User, PublicUser } from '~/app/types/auth'
+import { getUserById } from '~~/server/repositories/user.repository'
+import type { User, PublicUser } from '~~/shared/types/auth'
 import jwt from 'jsonwebtoken'
 
 const JWT_SECRET = process.env.JWT_SECRET || 'your-secret-key'
