+++ shared/types/auth.ts
